requirements.txt
djangopui
"djangorestframework==3.11.0" #NOT MAINTAINED
dj-rest-auth
"dj-rest-auth[with_social]"
djangorestframework-simplejwt
drf-yasg #TODO Check for rest_framework_swagger??
django-cors-headers

python-dotenv
Python 3.9.1 